add checker for "purity" of text
When it sees text that it doesn't consider to be part of the language's graphemes,
epitran passes that text over and lets it go into the transliteration as non IPA text.
When it doesn't need to do this, we can say the word is "pure".
It's not trivial to check whether a word is pure using the methods avaialbale in epitran,
due to preprocessing steps introducing more possible ipa segments than just what
appears as keys in the grapheme-to-phoneme mapping. Hence the introduction of this checker.